+++ ui/src/pages/nodes/view-inventory.ts
@@ -1,24 +1,10 @@
 // Nodes page renders the unified paired-device / node inventory card.
 import { html, nothing, type TemplateResult } from "lit";
-import {
-  GATEWAY_CLIENT_IDS,
-  GATEWAY_CLIENT_MODES,
-} from "../../../../packages/gateway-protocol/src/client-info.js";
-import {
-  resolvePendingDeviceApprovalState,
-  type DevicePairingAccessSummary,
-  type PendingDeviceApprovalKind,
-} from "../../../../src/shared/device-pairing-access.js";
 import type { PresenceEntry } from "../../api/types.ts";
 import { icons } from "../../components/icons.ts";
 import { t } from "../../i18n/index.ts";
 import { formatList, formatRelativeTimestamp, formatTimeAgo } from "../../lib/format.ts";
-import type {
-  DeviceTokenSummary,
-  InventoryRemovalRequest,
-  PairedDevice,
-  PendingDevice,
-} from "../../lib/nodes/index.ts";
+import type { DeviceTokenSummary, InventoryRemovalRequest } from "../../lib/nodes/index.ts";
 import {
   buildNodesInventory,
   findGatewayPresence,
@@ -29,6 +15,8 @@ import {
   type NodesInventoryGroup,
 } from "../../lib/nodes/inventory.ts";
 import { normalizeOptionalString } from "../../lib/string-coerce.ts";
+import { renderPendingDeviceRows } from "./view-pending-devices.ts";
+import { deviceIcon, renderDeviceTile, renderSectionLabel } from "./view-shared.ts";
 import type { NodesProps } from "./view.types.ts";
 
 const MAX_CAPABILITY_CHIPS = 16;
@@ -36,71 +24,6 @@ const MAX_CAPABILITY_CHIPS = 16;
 function toRemovalRequest(entry: NodesInventoryEntry): InventoryRemovalRequest {
   const removal = resolveInventoryRemoval(entry);
   return { id: entry.id, name: entry.name, ...removal };
-}
-
-type DeviceIconSource = {
-  clientId?: string;
-  clientMode?: string;
-  platform?: string;
-};
-
-const MOBILE_PLATFORM_PATTERN = /\b(ios|ipados|watchos|android|iphone|ipad)\b/;
-const MOBILE_CLIENT_IDS: ReadonlySet<string> = new Set([
-  GATEWAY_CLIENT_IDS.IOS_APP,
-  GATEWAY_CLIENT_IDS.WATCHOS_APP,
-  GATEWAY_CLIENT_IDS.ANDROID_APP,
-]);
-const BROWSER_CLIENT_IDS: ReadonlySet<string> = new Set([
-  GATEWAY_CLIENT_IDS.CONTROL_UI,
-  GATEWAY_CLIENT_IDS.WEBCHAT_UI,
-  GATEWAY_CLIENT_IDS.WEBCHAT,
-]);
-const TERMINAL_CLIENT_MODES: ReadonlySet<string> = new Set([
-  GATEWAY_CLIENT_MODES.CLI,
-  GATEWAY_CLIENT_MODES.BACKEND,
-  GATEWAY_CLIENT_MODES.PROBE,
-  GATEWAY_CLIENT_MODES.TEST,
-]);
-
-/** Rough form-factor icon: phone, browser, terminal, or desktop machine. */
-function deviceIcon(source: DeviceIconSource): TemplateResult {
-  const platform = source.platform?.trim().toLowerCase() ?? "";
-  const clientId = source.clientId?.trim().toLowerCase() ?? "";
-  const mode = source.clientMode?.trim().toLowerCase() ?? "";
-  if (MOBILE_PLATFORM_PATTERN.test(platform) || MOBILE_CLIENT_IDS.has(clientId)) {
-    return icons.smartphone;
-  }
-  if (BROWSER_CLIENT_IDS.has(clientId) || mode === GATEWAY_CLIENT_MODES.WEBCHAT) {
-    return icons.globe;
-  }
-  if (TERMINAL_CLIENT_MODES.has(mode) || clientId === GATEWAY_CLIENT_IDS.CLI) {
-    return icons.terminal;
-  }
-  return icons.monitor;
-}
-
-function renderDeviceTile(icon: TemplateResult, connected: boolean | null) {
-  return html`
-    <div class="nodes-entry__tile">
-      <span class="nodes-entry__tile-icon" aria-hidden="true">${icon}</span>
-      ${connected === null
-        ? nothing
-        : html`
-            <span
-              class="status-dot ${connected ? "status-dot--connected" : "status-dot--offline"}"
-              role="img"
-              aria-label=${connected
-                ? t("nodes.inventory.connected")
-                : t("nodes.inventory.offline")}
-              title=${connected ? t("nodes.inventory.connected") : t("nodes.inventory.offline")}
-            ></span>
-          `}
-    </div>
-  `;
-}
-
-function renderSectionLabel(label: string) {
-  return html`<div class="nodes-section-label">${label}</div>`;
 }
 
 function inventorySummary(
@@ -132,11 +55,6 @@ export function renderNodesInventory(props: NodesProps) {
   const gatewayPresence = findGatewayPresence(props.presence);
   const unpairedPresence = listUnpairedPresence(props.presence, groups);
   const stale = listStaleInventoryEntries(groups);
-  const pairedByDeviceId = new Map(
-    paired
-      .map((device) => [normalizeOptionalString(device.deviceId), device] as const)
-      .filter((entry): entry is [string, PairedDevice] => Boolean(entry[0])),
-  );
   const loading = props.loading || props.devicesLoading;
   return html`
     <section class="card">
@@ -177,9 +95,7 @@ export function renderNodesInventory(props: NodesProps) {
         ${pending.length > 0
           ? html`
               ${renderSectionLabel(t("nodes.inventory.pendingApproval"))}
-              ${pending.map((req) =>
-                renderPendingDevice(req, props, lookupPairedDevice(pairedByDeviceId, req)),
-              )}
+              ${renderPendingDeviceRows(pending, paired, props)}
               ${renderSectionLabel(t("nodes.inventory.paired"))}
             `
           : nothing}
@@ -527,99 +443,6 @@ function renderTokenRow(deviceId: string, token: DeviceTokenSummary, props: Node
                 ${t("nodes.inventory.revoke")}
               </button>
             `}
-      </div>
-    </div>
-  `;
-}
-
-function lookupPairedDevice(
-  pairedByDeviceId: ReadonlyMap<string, PairedDevice>,
-  request: Pick<PendingDevice, "deviceId" | "publicKey">,
-): PairedDevice | undefined {
-  const deviceId = normalizeOptionalString(request.deviceId);
-  if (!deviceId) {
-    return undefined;
-  }
-  const paired = pairedByDeviceId.get(deviceId);
-  if (!paired) {
-    return undefined;
-  }
-  const requestPublicKey = normalizeOptionalString(request.publicKey);
-  const pairedPublicKey = normalizeOptionalString(paired.publicKey);
-  if (requestPublicKey && pairedPublicKey && requestPublicKey !== pairedPublicKey) {
-    return undefined;
-  }
-  return paired;
-}
-
-function formatAccessSummary(access: DevicePairingAccessSummary | null): string {
-  if (!access) {
-    return t("nodes.inventory.none");
-  }
-  return t("nodes.inventory.rolesAndScopes", {
-    roles: formatList(access.roles),
-    scopes: formatList(access.scopes),
-  });
-}
-
-function renderPendingApprovalNote(kind: PendingDeviceApprovalKind) {
-  switch (kind) {
-    case "scope-upgrade":
-      return t("nodes.inventory.scopeUpgrade");
-    case "role-upgrade":
-      return t("nodes.inventory.roleUpgrade");
-    case "re-approval":
-      return t("nodes.inventory.reapproval");
-    case "new-pairing":
-      return t("nodes.inventory.newPairing");
-  }
-  const exhaustiveKind: never = kind;
-  void exhaustiveKind;
-  throw new Error("unsupported pending approval kind");
-}
-
-function renderPendingDevice(req: PendingDevice, props: NodesProps, paired?: PairedDevice) {
-  const name = normalizeOptionalString(req.displayName) || req.deviceId;
-  const age = typeof req.ts === "number" ? formatRelativeTimestamp(req.ts) : t("common.na");
-  const approval = resolvePendingDeviceApprovalState(req, paired);
-  const repair = req.isRepair ? ` · ${t("nodes.inventory.repair")}` : "";
-  const ip = req.remoteIp ? ` · ${req.remoteIp}` : "";
-  return html`
-    <div class="list-item nodes-entry nodes-entry--pending">
-      ${renderDeviceTile(icons.monitorSmartphone, null)}
-      <div class="list-main">
-        <div class="list-title">${name}</div>
-        <div class="list-sub">${req.deviceId}${ip}</div>
-        <div class="muted" style="margin-top: 6px;">
-          ${t("nodes.inventory.requestedAt", {
-            note: renderPendingApprovalNote(approval.kind),
-            time: age,
-          })}${repair}
-        </div>
-        <div class="muted" style="margin-top: 6px;">
-          ${t("nodes.inventory.requestedAccess", {
-            access: formatAccessSummary(approval.requested),
-          })}
-        </div>
-        ${approval.approved
-          ? html`
-              <div class="muted" style="margin-top: 6px;">
-                ${t("nodes.inventory.approvedAccess", {
-                  access: formatAccessSummary(approval.approved),
-                })}
-              </div>
-            `
-          : nothing}
-      </div>
-      <div class="list-meta">
-        <div class="row" style="justify-content: flex-end; gap: 8px; flex-wrap: wrap;">
-          <button class="btn btn--sm primary" @click=${() => props.onDeviceApprove(req.requestId)}>
-            ${t("nodes.inventory.approve")}
-          </button>
-          <button class="btn btn--sm" @click=${() => props.onDeviceReject(req.requestId)}>
-            ${t("nodes.inventory.reject")}
-          </button>
-        </div>
       </div>
     </div>
   `;
