+++ ui/src/pages/nodes/view-shared.ts
@@ -1,4 +1,11 @@
 // Nodes page owns these pure view helpers.
+import { html, nothing, type TemplateResult } from "lit";
+import {
+  GATEWAY_CLIENT_IDS,
+  GATEWAY_CLIENT_MODES,
+} from "../../../../packages/gateway-protocol/src/client-info.js";
+import { icons } from "../../components/icons.ts";
+import { t } from "../../i18n/index.ts";
 import { normalizeOptionalString } from "../../lib/string-coerce.ts";
 
 export type NodeTargetOption = {
@@ -63,4 +70,70 @@ export function resolveNodeTargets(
 
   list.sort((a, b) => a.label.localeCompare(b.label));
   return list;
+}
+
+export type DeviceIconSource = {
+  clientId?: string;
+  clientMode?: string;
+  platform?: string;
+};
+
+const MOBILE_PLATFORM_PATTERN = /\b(ios|ipados|watchos|android|iphone|ipad)\b/;
+const MOBILE_CLIENT_IDS: ReadonlySet<string> = new Set([
+  GATEWAY_CLIENT_IDS.IOS_APP,
+  GATEWAY_CLIENT_IDS.WATCHOS_APP,
+  GATEWAY_CLIENT_IDS.ANDROID_APP,
+]);
+const BROWSER_CLIENT_IDS: ReadonlySet<string> = new Set([
+  GATEWAY_CLIENT_IDS.CONTROL_UI,
+  GATEWAY_CLIENT_IDS.WEBCHAT_UI,
+  GATEWAY_CLIENT_IDS.WEBCHAT,
+]);
+const TERMINAL_CLIENT_MODES: ReadonlySet<string> = new Set([
+  GATEWAY_CLIENT_MODES.CLI,
+  GATEWAY_CLIENT_MODES.BACKEND,
+  GATEWAY_CLIENT_MODES.PROBE,
+  GATEWAY_CLIENT_MODES.TEST,
+]);
+
+/** Rough form-factor icon: phone, browser, terminal, or desktop machine. */
+export function deviceIcon(source: DeviceIconSource): TemplateResult {
+  const platform = source.platform?.trim().toLowerCase() ?? "";
+  const clientId = source.clientId?.trim().toLowerCase() ?? "";
+  const mode = source.clientMode?.trim().toLowerCase() ?? "";
+  if (MOBILE_PLATFORM_PATTERN.test(platform) || MOBILE_CLIENT_IDS.has(clientId)) {
+    return icons.smartphone;
+  }
+  if (BROWSER_CLIENT_IDS.has(clientId) || mode === GATEWAY_CLIENT_MODES.WEBCHAT) {
+    return icons.globe;
+  }
+  if (TERMINAL_CLIENT_MODES.has(mode) || clientId === GATEWAY_CLIENT_IDS.CLI) {
+    return icons.terminal;
+  }
+  return icons.monitor;
+}
+
+/** Icon tile with a connectivity dot; `connected: null` hides the dot. */
+export function renderDeviceTile(icon: TemplateResult, connected: boolean | null) {
+  return html`
+    <div class="nodes-entry__tile">
+      <span class="nodes-entry__tile-icon" aria-hidden="true">${icon}</span>
+      ${connected === null
+        ? nothing
+        : html`
+            <span
+              class="status-dot ${connected ? "status-dot--connected" : "status-dot--offline"}"
+              role="img"
+              aria-label=${connected
+                ? t("nodes.inventory.connected")
+                : t("nodes.inventory.offline")}
+              title=${connected ? t("nodes.inventory.connected") : t("nodes.inventory.offline")}
+            ></span>
+          `}
+    </div>
+  `;
+}
+
+export function renderSectionLabel(label: string) {
+  return html`<div class="nodes-section-label">${label}</div>`;
 }
